refactor(python): split auth base forwarder tests by contract
