+++ crates/runner/mitm-addon/tests/test_auth_base_forwarder_lifecycle.py
@@ -1,16 +1,13 @@
-"""Tests for auth.base HTTPS forwarding behavior."""
+"""Admission, deadline, worker, and shutdown tests for auth.base forwarding."""
 
 import asyncio
 import contextvars
-import errno
-import http.client as http_client
 import multiprocessing
-import ssl
 import threading
 import time
 from collections.abc import Callable
 from concurrent.futures import Future
-from unittest.mock import MagicMock, call, patch
+from unittest.mock import MagicMock, patch
 
 import pytest
 from mitmproxy import http
@@ -93,947 +90,6 @@ def _run_blocked_forward_then_shutdown_wait_false() -> None:
                 task.cancel()
 
     asyncio.run(main())
-
-
-class TestAuthBaseForwarderSecurity:
-    @pytest.mark.parametrize(
-        ("url", "resolved_address"),
-        [
-            pytest.param("https://127.0.0.1/", "127.0.0.1", id="ipv4-loopback"),
-            pytest.param(
-                "https://169.254.169.254/",
-                "169.254.169.254",
-                id="ipv4-link-local-metadata",
-            ),
-            pytest.param("https://10.0.0.1/", "10.0.0.1", id="ipv4-rfc1918-10"),
-            pytest.param("https://172.16.0.1/", "172.16.0.1", id="ipv4-rfc1918-172"),
-            pytest.param("https://192.168.0.1/", "192.168.0.1", id="ipv4-rfc1918-192"),
-            pytest.param("https://100.64.0.1/", "100.64.0.1", id="ipv4-cgnat"),
-            pytest.param("https://224.0.0.1/", "224.0.0.1", id="ipv4-multicast"),
-            pytest.param("https://240.0.0.1/", "240.0.0.1", id="ipv4-reserved"),
-            pytest.param("https://[::1]/", "::1", id="ipv6-loopback"),
-            pytest.param("https://[fe80::1]/", "fe80::1", id="ipv6-link-local"),
-            pytest.param("https://[fc00::1]/", "fc00::1", id="ipv6-ula"),
-            pytest.param("https://[2001:db8::1]/", "2001:db8::1", id="ipv6-documentation"),
-            pytest.param("https://[3fff::1]/", "3fff::1", id="ipv6-expanded-documentation"),
-            pytest.param(
-                "https://[::ffff:100.64.0.1]/",
-                "::ffff:100.64.0.1",
-                id="ipv6-mapped-cgnat",
-            ),
-            pytest.param(
-                "https://[::ffff:8.8.8.8]/",
-                "::ffff:8.8.8.8",
-                id="ipv6-mapped-reserved",
-            ),
-            pytest.param("https://[2002:0808:0808::1]/", "2002:0808:0808::1", id="ipv6-6to4"),
-            pytest.param(
-                "https://[2001:0000:4136:e378:8000:63bf:3fff:fdd2]/",
-                "2001:0000:4136:e378:8000:63bf:3fff:fdd2",
-                id="ipv6-teredo",
-            ),
-            pytest.param(
-                "https://[64:ff9b::169.254.169.254]/",
-                "64:ff9b::169.254.169.254",
-                id="ipv6-nat64-metadata",
-            ),
-            pytest.param(
-                "https://[64:ff9b::8.8.8.8]/",
-                "64:ff9b::8.8.8.8",
-                id="ipv6-nat64-reserved",
-            ),
-        ],
-    )
-    async def test_rejects_non_public_destinations_without_opening_connection(
-        self,
-        url: str,
-        resolved_address: str,
-    ):
-        with (
-            fake_forwarder_upstream(addresses=(resolved_address,)) as upstream,
-            pytest.raises(
-                forwarder.UnsafeAuthBaseDestinationError,
-                match=r"Unsafe auth\.base upstream destination",
-            ),
-        ):
-            await forwarder.forward_request(url, "GET", [], None)
-
-        assert upstream.connect_calls == []
-        assert upstream.sockets == []
-
-    async def test_rejects_dns_private_destination_without_opening_connection(self):
-        with (
-            fake_forwarder_upstream(addresses=("10.0.0.1",)) as upstream,
-            pytest.raises(forwarder.UnsafeAuthBaseDestinationError),
-        ):
-            await forwarder.forward_request("https://hooks.example.com/path", "GET", [], None)
-
-        assert upstream.resolve_calls == ["hooks.example.com"]
-        assert upstream.connect_calls == []
-
-    async def test_rejects_mixed_dns_answers_without_opening_connection(self):
-        with (
-            fake_forwarder_upstream(addresses=("93.184.216.34", "127.0.0.1")) as upstream,
-            pytest.raises(forwarder.UnsafeAuthBaseDestinationError),
-        ):
-            await forwarder.forward_request("https://hooks.example.com/path", "GET", [], None)
-
-        assert upstream.resolve_calls == ["hooks.example.com"]
-        assert upstream.connect_calls == []
-
-    async def test_allows_public_dns_destination_and_forwards_with_original_host(self):
-        with fake_forwarder_upstream(
-            addresses=("93.184.216.34", "2001:1::3", "2001:4860:4860::8888")
-        ) as upstream:
-            status, body, headers = await forwarder.forward_request(
-                "https://hooks.example.com/path",
-                "GET",
-                [],
-                None,
-            )
-
-        assert status == 200
-        assert body == b"ok"
-        assert list(headers.items(multi=True)) == []
-        assert upstream.resolve_calls == ["hooks.example.com"]
-        assert upstream.connect_calls == [("93.184.216.34", 443)]
-        assert upstream.contexts[-1].server_hostnames == ["hooks.example.com"]
-        assert upstream.socket.request_lines()[0] == "GET /path HTTP/1.1"
-        assert upstream.socket.request_header_values("Host") == ["hooks.example.com"]
-
-    async def test_reuses_https_context_without_reusing_connections(self):
-        with fake_forwarder_upstream() as upstream:
-            first_status, first_body, _first_headers = await forwarder.forward_request(
-                "https://example.com/one",
-                "GET",
-                [],
-                None,
-            )
-            second_status, second_body, _second_headers = await forwarder.forward_request(
-                "https://example.com/two",
-                "GET",
-                [],
-                None,
-            )
-
-        assert first_status == 200
-        assert first_body == b"ok"
-        assert second_status == 200
-        assert second_body == b"ok"
-        assert len(upstream.contexts) == 1
-        assert upstream.contexts[0].server_hostnames == ["example.com", "example.com"]
-        assert upstream.resolve_calls == ["example.com", "example.com"]
-        assert upstream.connect_calls == [
-            ("93.184.216.34", 443),
-            ("93.184.216.34", 443),
-        ]
-        assert [socket.request_lines()[0] for socket in upstream.sockets] == [
-            "GET /one HTTP/1.1",
-            "GET /two HTTP/1.1",
-        ]
-        assert [socket.closed for socket in upstream.sockets] == [True, True]
-
-    @pytest.mark.parametrize(
-        "url",
-        [
-            pytest.param("file:///etc/passwd", id="file"),
-            pytest.param("ftp://evil.com/file", id="ftp"),
-            pytest.param("http://example.com/path", id="http"),
-            pytest.param("//no-scheme.com/path", id="empty-scheme"),
-        ],
-    )
-    async def test_rejects_unsupported_scheme_before_dns(self, url: str):
-        with (
-            fake_forwarder_upstream() as upstream,
-            pytest.raises(ValueError, match="Unsupported URL scheme"),
-        ):
-            await forwarder.forward_request(url, "GET", [], None)
-
-        assert upstream.resolve_calls == []
-
-    async def test_rejects_missing_host_before_dns(self):
-        with (
-            fake_forwarder_upstream() as upstream,
-            pytest.raises(ValueError, match="Invalid upstream URL: missing host"),
-        ):
-            await forwarder.forward_request("https:///path", "GET", [], None)
-
-        assert upstream.resolve_calls == []
-
-    @pytest.mark.parametrize(
-        "url",
-        [
-            "https://example.com:bad/path",
-            "https://example.com:/path",
-        ],
-    )
-    async def test_rejects_invalid_port_before_dns(self, url: str):
-        with (
-            fake_forwarder_upstream() as upstream,
-            pytest.raises(ValueError, match="Invalid upstream URL: invalid port"),
-        ):
-            await forwarder.forward_request(url, "GET", [], None)
-
-        assert upstream.resolve_calls == []
-
-    async def test_rejects_bracketed_non_ipv6_authority_before_dns(self):
-        with (
-            fake_forwarder_upstream() as upstream,
-            pytest.raises(ValueError, match="Invalid upstream URL: invalid host"),
-        ):
-            await forwarder.forward_request("https://[v1.invalid]/path", "GET", [], None)
-
-        assert upstream.resolve_calls == []
-
-    @pytest.mark.parametrize(
-        "url",
-        [
-            "https://fa\u212a.example/path",
-            "https://\u212a.example/path",
-            "https://%E2%84%AA.example/path",
-            "https://example%2ecom/path",
-            "https://*.example.com/path",
-            "https://api*.example.com/path",
-            "https://{env}.example.com/path",
-            "https://api{env}.example.com/path",
-            "https://%2A.example.com/path",
-        ],
-    )
-    async def test_rejects_unsafe_raw_host_before_dns(self, url: str):
-        with (
-            fake_forwarder_upstream() as upstream,
-            pytest.raises(ValueError, match="Invalid upstream URL: invalid host"),
-        ):
-            await forwarder.forward_request(url, "GET", [], None)
-
-        assert upstream.resolve_calls == []
-
-    @pytest.mark.parametrize(
-        ("url", "expected_host"),
-        [
-            ("https://b\u00fccher.example/path", "xn--bcher-kva.example"),
-            ("https://b%C3%BCcher.example/path", "xn--bcher-kva.example"),
-            ("https://fa\u00df.example/path", "xn--fa-hia.example"),
-        ],
-    )
-    async def test_normalizes_idna_host_before_forwarding(self, url: str, expected_host: str):
-        with fake_forwarder_upstream() as upstream:
-            await forwarder.forward_request(url, "GET", [], None)
-
-        assert upstream.resolve_calls == [expected_host]
-        assert upstream.contexts[-1].server_hostnames == [expected_host]
-        assert upstream.socket.request_header_values("Host") == [expected_host]
-
-    @pytest.mark.parametrize(
-        "url",
-        [
-            "https://user@example.com/path",
-            "https://user:pass@example.com/path",
-        ],
-    )
-    async def test_rejects_userinfo_authority_before_dns(self, url: str):
-        with (
-            fake_forwarder_upstream() as upstream,
-            pytest.raises(ValueError, match="Unsupported URL authority"),
-        ):
-            await forwarder.forward_request(url, "GET", [], None)
-
-        assert upstream.resolve_calls == []
-
-
-class TestAuthBaseForwarderTransportSecurity:
-    def test_validated_connection_uses_checked_ip_and_original_hostname_for_sni(self):
-        raw_sock = MagicMock()
-        wrapped_sock = MagicMock()
-        context = MagicMock()
-        context.wrap_socket.return_value = wrapped_sock
-        abort_handle = forwarder._ForwardRequestAbortHandle(MagicMock())
-        conn = forwarder._make_validated_https_connection(
-            "hooks.example.com",
-            port=None,
-            deadline=time.monotonic() + 30,
-            abort_handle=abort_handle,
-            validated_addresses=(
-                forwarder._ValidatedAddress(
-                    forwarder.socket.AF_INET,
-                    "93.184.216.34",
-                    443,
-                ),
-            ),
-        )
-        vars(conn)["_context"] = context
-
-        with patch.object(forwarder.socket, "socket", return_value=raw_sock) as create_socket:
-            conn.connect()
-
-        create_socket.assert_called_once_with(
-            forwarder.socket.AF_INET,
-            forwarder.socket.SOCK_STREAM,
-        )
-        raw_sock.connect.assert_called_once_with(("93.184.216.34", 443))
-        raw_sock.setsockopt.assert_called_once_with(
-            forwarder.socket.IPPROTO_TCP,
-            forwarder.socket.TCP_NODELAY,
-            1,
-        )
-        context.wrap_socket.assert_called_once_with(
-            raw_sock,
-            server_hostname="hooks.example.com",
-            do_handshake_on_connect=False,
-        )
-        wrapped_sock.do_handshake.assert_called_once_with()
-        assert conn.sock is wrapped_sock
-
-    def test_validated_connection_retries_checked_addresses_without_new_dns(self):
-        failed_sock = MagicMock()
-        failed_sock.connect.side_effect = OSError("no route")
-        raw_sock = MagicMock()
-        wrapped_sock = MagicMock()
-        context = MagicMock()
-        context.wrap_socket.return_value = wrapped_sock
-        abort_handle = forwarder._ForwardRequestAbortHandle(MagicMock())
-        conn = forwarder._make_validated_https_connection(
-            "hooks.example.com",
-            port=None,
-            deadline=time.monotonic() + 30,
-            abort_handle=abort_handle,
-            validated_addresses=(
-                forwarder._ValidatedAddress(
-                    forwarder.socket.AF_INET6,
-                    "2001:4860:4860::8888",
-                    443,
-                ),
-                forwarder._ValidatedAddress(
-                    forwarder.socket.AF_INET,
-                    "93.184.216.34",
-                    443,
-                ),
-            ),
-        )
-        vars(conn)["_context"] = context
-
-        with patch.object(
-            forwarder.socket,
-            "socket",
-            side_effect=[failed_sock, raw_sock],
-        ) as create_socket:
-            conn.connect()
-
-        create_socket.assert_has_calls(
-            [
-                call(forwarder.socket.AF_INET6, forwarder.socket.SOCK_STREAM),
-                call(forwarder.socket.AF_INET, forwarder.socket.SOCK_STREAM),
-            ]
-        )
-        failed_sock.connect.assert_called_once_with(("2001:4860:4860::8888", 443, 0, 0))
-        failed_sock.close.assert_called_once_with()
-        raw_sock.connect.assert_called_once_with(("93.184.216.34", 443))
-        raw_sock.setsockopt.assert_called_once_with(
-            forwarder.socket.IPPROTO_TCP,
-            forwarder.socket.TCP_NODELAY,
-            1,
-        )
-        context.wrap_socket.assert_called_once_with(
-            raw_sock,
-            server_hostname="hooks.example.com",
-            do_handshake_on_connect=False,
-        )
-        wrapped_sock.do_handshake.assert_called_once_with()
-        assert conn.sock is wrapped_sock
-
-    def test_validated_connection_context_keeps_https_security_defaults(self):
-        context = forwarder._create_https_context()
-
-        assert context.verify_mode is ssl.CERT_REQUIRED
-        assert context.check_hostname is True
-        if context.post_handshake_auth is not None:
-            assert context.post_handshake_auth is True
-
-    def test_validated_connection_ignores_missing_tcp_nodelay(self):
-        raw_sock = MagicMock()
-        raw_sock.setsockopt.side_effect = OSError(errno.ENOPROTOOPT, "not supported")
-        wrapped_sock = MagicMock()
-        context = MagicMock()
-        context.wrap_socket.return_value = wrapped_sock
-        abort_handle = forwarder._ForwardRequestAbortHandle(MagicMock())
-        conn = forwarder._make_validated_https_connection(
-            "hooks.example.com",
-            port=None,
-            deadline=time.monotonic() + 30,
-            abort_handle=abort_handle,
-            validated_addresses=(
-                forwarder._ValidatedAddress(
-                    forwarder.socket.AF_INET,
-                    "93.184.216.34",
-                    443,
-                ),
-            ),
-        )
-        vars(conn)["_context"] = context
-
-        with patch.object(forwarder.socket, "socket", return_value=raw_sock):
-            conn.connect()
-
-        raw_sock.close.assert_not_called()
-        context.wrap_socket.assert_called_once_with(
-            raw_sock,
-            server_hostname="hooks.example.com",
-            do_handshake_on_connect=False,
-        )
-        assert conn.sock is wrapped_sock
-
-    def test_validated_connection_clears_raw_socket_when_setsockopt_cleanup_close_fails(
-        self,
-    ):
-        raw_sock = MagicMock()
-        raw_sock.setsockopt.side_effect = OSError(errno.EINVAL, "setsockopt failed")
-        raw_sock.close.side_effect = OSError("close failed")
-        context = MagicMock()
-        abort_handle = forwarder._ForwardRequestAbortHandle(MagicMock())
-        conn = forwarder._make_validated_https_connection(
-            "hooks.example.com",
-            port=None,
-            deadline=time.monotonic() + 30,
-            abort_handle=abort_handle,
-            validated_addresses=(
-                forwarder._ValidatedAddress(
-                    forwarder.socket.AF_INET,
-                    "93.184.216.34",
-                    443,
-                ),
-            ),
-        )
-        vars(conn)["_context"] = context
-
-        with (
-            patch.object(forwarder.socket, "socket", return_value=raw_sock),
-            pytest.raises(OSError, match="setsockopt failed"),
-        ):
-            conn.connect()
-
-        raw_sock.close.assert_called_once_with()
-        context.wrap_socket.assert_not_called()
-        assert abort_handle.abort_for_shutdown()
-        raw_sock.shutdown.assert_not_called()
-
-    def test_validated_connection_closes_raw_socket_when_tls_wrap_fails(self):
-        raw_sock = MagicMock()
-        context = MagicMock()
-        context.wrap_socket.side_effect = OSError("tls failed")
-        abort_handle = forwarder._ForwardRequestAbortHandle(MagicMock())
-        conn = forwarder._make_validated_https_connection(
-            "hooks.example.com",
-            port=None,
-            deadline=time.monotonic() + 30,
-            abort_handle=abort_handle,
-            validated_addresses=(
-                forwarder._ValidatedAddress(
-                    forwarder.socket.AF_INET,
-                    "93.184.216.34",
-                    443,
-                ),
-            ),
-        )
-        vars(conn)["_context"] = context
-
-        with (
-            patch.object(forwarder.socket, "socket", return_value=raw_sock),
-            pytest.raises(OSError, match="tls failed"),
-        ):
-            conn.connect()
-
-        raw_sock.close.assert_called_once_with()
-
-    def test_validated_connection_clears_raw_socket_when_tls_cleanup_close_fails(self):
-        raw_sock = MagicMock()
-        raw_sock.close.side_effect = OSError("close failed")
-        context = MagicMock()
-        context.wrap_socket.side_effect = OSError("tls failed")
-        abort_handle = forwarder._ForwardRequestAbortHandle(MagicMock())
-        conn = forwarder._make_validated_https_connection(
-            "hooks.example.com",
-            port=None,
-            deadline=time.monotonic() + 30,
-            abort_handle=abort_handle,
-            validated_addresses=(
-                forwarder._ValidatedAddress(
-                    forwarder.socket.AF_INET,
-                    "93.184.216.34",
-                    443,
-                ),
-            ),
-        )
-        vars(conn)["_context"] = context
-
-        with (
-            patch.object(forwarder.socket, "socket", return_value=raw_sock),
-            pytest.raises(OSError, match="tls failed"),
-        ):
-            conn.connect()
-
-        raw_sock.close.assert_called_once_with()
-        assert abort_handle.abort_for_shutdown()
-        raw_sock.shutdown.assert_not_called()
-
-
-class TestAuthBaseForwarderRequestBehavior:
-    async def test_returns_redirect_response_without_following(self):
-        with fake_forwarder_upstream(
-            status=302,
-            body=b"",
-            headers=[("Location", "https://evil.example.com")],
-        ):
-            status, body, headers = await forwarder.forward_request(
-                "https://example.com/redirect",
-                "GET",
-                [],
-                None,
-            )
-
-        assert status == 302
-        assert body == b""
-        assert headers["Location"] == "https://evil.example.com"
-
-    async def test_repeated_request_headers_are_written_individually(self):
-        with fake_forwarder_upstream() as upstream:
-            await forwarder.forward_request(
-                "https://example.com/path?x=1",
-                "GET",
-                [("X-Repeat", "one"), ("X-Repeat", "two")],
-                None,
-            )
-
-        assert upstream.socket.request_lines()[0] == "GET /path?x=1 HTTP/1.1"
-        assert upstream.socket.request_header_values("Host") == ["example.com"]
-        assert upstream.socket.request_header_values("X-Repeat") == ["one", "two"]
-        assert upstream.socket.request_header_values("Content-Length") == []
-
-    async def test_absent_body_strips_stale_content_length(self):
-        with fake_forwarder_upstream() as upstream:
-            await forwarder.forward_request(
-                "https://example.com/path",
-                "POST",
-                [("Content-Length", "999"), ("X-Keep", "ok")],
-                None,
-            )
-
-        assert upstream.socket.request_header_values("Content-Length") == []
-        assert upstream.socket.request_header_values("X-Keep") == ["ok"]
-        assert upstream.socket.request_text().endswith("\r\n\r\n")
-
-    @pytest.mark.parametrize(
-        ("url", "expected_target"),
-        [
-            pytest.param(
-                "https://example.com?wait=true",
-                "/?wait=true",
-                id="root-query",
-            ),
-            pytest.param(
-                "https://example.com/path?x=1#client-only-secret",
-                "/path?x=1",
-                id="omit-fragment",
-            ),
-            pytest.param(
-                "https://example.com/%2Fsecret/a%20b?x=a%2Fb&x=&space=a+b",
-                "/%2Fsecret/a%20b?x=a%2Fb&x=&space=a+b",
-                id="encoded-path-duplicate-query",
-            ),
-            pytest.param(
-                "https://example.com/hook;v=1/sub;mode=fast?x=1",
-                "/hook;v=1/sub;mode=fast?x=1",
-                id="path-params",
-            ),
-        ],
-    )
-    async def test_request_target_preserves_url_parts(self, url: str, expected_target: str):
-        with fake_forwarder_upstream() as upstream:
-            await forwarder.forward_request(url, "GET", [], None)
-
-        assert upstream.socket.request_lines()[0] == f"GET {expected_target} HTTP/1.1"
-
-    @pytest.mark.parametrize(
-        (
-            "url",
-            "expected_dns_host",
-            "expected_connection_port",
-            "expected_host_header",
-        ),
-        [
-            pytest.param(
-                "https://example.com:443/path",
-                "example.com",
-                443,
-                "example.com",
-                id="https-default-port",
-            ),
-            pytest.param(
-                "https://[2001:4860:4860::8888]:444/path",
-                "2001:4860:4860::8888",
-                444,
-                "[2001:4860:4860::8888]:444",
-                id="ipv6-non-default-port",
-            ),
-            pytest.param(
-                "https://[2001:4860:4860::8888]/path",
-                "2001:4860:4860::8888",
-                443,
-                "[2001:4860:4860::8888]",
-                id="ipv6-no-port",
-            ),
-            pytest.param(
-                "https://[2001:4860:4860::8888]:443/path",
-                "2001:4860:4860::8888",
-                443,
-                "[2001:4860:4860::8888]",
-                id="ipv6-https-default-port",
-            ),
-            pytest.param(
-                "https://[2001:4860:4860::8888]:80/path",
-                "2001:4860:4860::8888",
-                80,
-                "[2001:4860:4860::8888]:80",
-                id="ipv6-https-http-default-port",
-            ),
-        ],
-    )
-    async def test_url_authority_sets_connection_target_and_host_header(
-        self,
-        url: str,
-        expected_dns_host: str,
-        expected_connection_port: int,
-        expected_host_header: str,
-    ):
-        with fake_forwarder_upstream() as upstream:
-            await forwarder.forward_request(
-                url,
-                "GET",
-                [],
-                None,
-            )
-
-        expected_resolve_calls = [] if ":" in expected_dns_host else [expected_dns_host]
-        assert upstream.resolve_calls == expected_resolve_calls
-        expected_connect_address = (
-            (expected_dns_host, expected_connection_port, 0, 0)
-            if ":" in expected_dns_host
-            else ("93.184.216.34", expected_connection_port)
-        )
-        assert upstream.connect_calls == [expected_connect_address]
-        assert upstream.socket.request_header_values("Host") == [expected_host_header]
-
-    async def test_filters_request_hop_by_hop_headers_and_recomputes_content_length(self):
-        with fake_forwarder_upstream() as upstream:
-            await forwarder.forward_request(
-                "https://example.com:444/path",
-                "PUT",
-                [
-                    ("Host", "agent.example.com"),
-                    ("Connection", "X-Remove, Keep-Alive"),
-                    ("X-Remove", "secret"),
-                    ("Keep-Alive", "timeout=5"),
-                    ("Proxy-Authorization", "Basic secret"),
-                    ("Content-Length", "999"),
-                    ("Transfer-Encoding", "chunked"),
-                    ("X-Keep", "ok"),
-                ],
-                b"abc",
-            )
-
-        assert upstream.socket.request_header_values("Connection") == []
-        assert upstream.socket.request_header_values("X-Remove") == []
-        assert upstream.socket.request_header_values("Keep-Alive") == []
-        assert upstream.socket.request_header_values("Proxy-Authorization") == []
-        assert upstream.socket.request_header_values("Transfer-Encoding") == []
-        assert upstream.socket.request_header_values("Host") == ["example.com:444"]
-        assert upstream.socket.request_header_values("X-Keep") == ["ok"]
-        assert upstream.socket.request_header_values("Content-Length") == ["3"]
-        assert upstream.socket.request_text().endswith("\r\n\r\nabc")
-
-    async def test_explicit_empty_body_sets_zero_content_length(self):
-        with fake_forwarder_upstream() as upstream:
-            await forwarder.forward_request(
-                "https://example.com/path",
-                "POST",
-                [],
-                b"",
-            )
-
-        assert upstream.socket.request_header_values("Content-Length") == ["0"]
-        assert upstream.socket.request_text().endswith("\r\n\r\n")
-
-    async def test_preserves_response_header_octets_duplicates_and_filters_connection_names(self):
-        raw_response = (
-            b"HTTP/1.1 200 OK\r\n"
-            b"X-Bytes: caf\xe9\r\n"
-            b"Set-Cookie: a=1\r\n"
-            b"Set-Cookie: b=2\r\n"
-            b"Connection: X-Remove\r\n"
-            b"X-Remove: drop\r\n"
-            b"X-Keep: ok\r\n"
-            b"\r\n"
-            b"ok"
-        )
-
-        with fake_forwarder_upstream(socket_factory=lambda: FakeSocket(raw_response)):
-            status, body, headers = await forwarder.forward_request(
-                "https://example.com",
-                "GET",
-                [],
-                None,
-            )
-
-        assert status == 200
-        assert body == b"ok"
-        assert headers.fields == (
-            (b"X-Bytes", b"caf\xe9"),
-            (b"Set-Cookie", b"a=1"),
-            (b"Set-Cookie", b"b=2"),
-            (b"X-Keep", b"ok"),
-        )
-
-    async def test_filters_hop_by_hop_response_headers(self):
-        with fake_forwarder_upstream(
-            body=b"2\r\nok\r\n0\r\n\r\n",
-            headers=[
-                ("Content-Type", "application/json"),
-                ("Transfer-Encoding", "chunked"),
-                ("Connection", "keep-alive"),
-                ("Proxy-Authenticate", "Basic realm=proxy"),
-                ("X-Custom", "value"),
-            ],
-        ):
-            _status, _body, headers = await forwarder.forward_request(
-                "https://example.com",
-                "GET",
-                [],
-                None,
-            )
-
-        assert _body == b"ok"
-        assert "Content-Type" in headers
-        assert "X-Custom" in headers
-        assert "Transfer-Encoding" not in headers
-        assert "Connection" not in headers
-        assert "Proxy-Authenticate" not in headers
-
-
-class TestAuthBaseForwarderResponseBodyLimit:
-    async def test_accepts_body_at_limit(self):
-        with (
-            patch.object(forwarder, "MAX_AUTH_BASE_RESPONSE_BODY_BYTES", 4),
-            fake_forwarder_upstream(body=b"1234") as upstream,
-        ):
-            status, body, _headers = await forwarder.forward_request(
-                "https://example.com",
-                "GET",
-                [],
-                None,
-            )
-
-        assert status == 200
-        assert body == b"1234"
-        assert upstream.socket.response_file is not None
-        assert upstream.socket.response_file.read_sizes == [5]
-
-    async def test_rejects_body_over_limit_and_closes_resources(self):
-        with (
-            patch.object(forwarder, "MAX_AUTH_BASE_RESPONSE_BODY_BYTES", 4),
-            fake_forwarder_upstream(body=b"12345") as upstream,
-            pytest.raises(forwarder.ForwardedResponseTooLargeError),
-        ):
-            await forwarder.forward_request("https://example.com", "GET", [], None)
-
-        assert upstream.socket.response_file is not None
-        assert upstream.socket.response_file.read_sizes == [5]
-        assert upstream.socket.response_file.closed
-        assert upstream.socket.closed
-
-
-class TestAuthBaseForwarderRequestBodyLimit:
-    async def test_accepts_body_at_limit(self):
-        with (
-            patch.object(forwarder, "MAX_AUTH_BASE_REQUEST_BODY_BYTES", 4),
-            fake_forwarder_upstream() as upstream,
-        ):
-            status, body, _headers = await forwarder.forward_request(
-                "https://example.com",
-                "POST",
-                [],
-                b"1234",
-            )
-
-        assert status == 200
-        assert body == b"ok"
-        assert upstream.socket.request_text().endswith("\r\n\r\n1234")
-
-    async def test_rejects_body_over_limit_before_connection_setup(self):
-        with (
-            patch.object(forwarder, "MAX_AUTH_BASE_REQUEST_BODY_BYTES", 4),
-            fake_forwarder_upstream() as upstream,
-            pytest.raises(forwarder.ForwardedRequestTooLargeError),
-        ):
-            await forwarder.forward_request(
-                "https://example.com",
-                "POST",
-                [],
-                b"12345",
-            )
-
-        assert upstream.resolve_calls == []
-
-
-class TestAuthBaseForwarderResourceCleanup:
-    def test_duplicate_flow_admission_attach_does_not_overwrite_existing(self, real_flow):
-        flow = real_flow(with_response=False)
-        first = forwarder.reserve_forward_request_admission(7)
-        second = forwarder.reserve_forward_request_admission(11)
-
-        try:
-            forwarder.attach_forward_request_admission_to_flow(flow, first)
-
-            with pytest.raises(RuntimeError, match="already attached"):
-                forwarder.attach_forward_request_admission_to_flow(flow, second)
-
-            assert forwarder.forward_request_admission_state_for_tests() == (2, 18)
-        finally:
-            forwarder.release_forward_request_admission_from_flow(flow)
-            forwarder.release_forward_request_admission(first)
-            forwarder.release_forward_request_admission(second)
-
-        assert forwarder.forward_request_admission_state_for_tests() == (0, 0)
-
-    async def test_oversized_body_releases_supplied_admission_before_upstream_setup(self):
-        with (
-            patch.object(forwarder, "MAX_AUTH_BASE_REQUEST_BODY_BYTES", 4),
-            patch.object(forwarder, "MAX_ADMITTED_AUTH_BASE_FORWARDS", 1),
-            patch.object(forwarder, "MAX_ADMITTED_AUTH_BASE_REQUEST_BODY_BYTES", 4),
-            fake_forwarder_upstream() as upstream,
-        ):
-            admission = forwarder.reserve_forward_request_admission(4)
-
-            with pytest.raises(forwarder.ForwardedRequestTooLargeError):
-                await forwarder.forward_request(
-                    "https://example.com",
-                    "POST",
-                    [],
-                    b"12345",
-                    admission=admission,
-                )
-
-            assert forwarder.forward_request_admission_state_for_tests() == (0, 0)
-            replacement = forwarder.reserve_forward_request_admission(4)
-            try:
-                assert forwarder.forward_request_admission_state_for_tests() == (1, 4)
-            finally:
-                forwarder.release_forward_request_admission(replacement)
-
-        assert forwarder.forward_request_admission_state_for_tests() == (0, 0)
-        assert upstream.resolve_calls == []
-        assert upstream.sockets == []
-        assert upstream.connect_calls == []
-
-    async def test_resize_saturation_releases_only_supplied_admission(self):
-        with (
-            patch.object(forwarder, "MAX_ADMITTED_AUTH_BASE_FORWARDS", 2),
-            patch.object(forwarder, "MAX_ADMITTED_AUTH_BASE_REQUEST_BODY_BYTES", 4),
-            fake_forwarder_upstream() as upstream,
-        ):
-            unrelated = forwarder.reserve_forward_request_admission(3)
-            admission = forwarder.reserve_forward_request_admission(1)
-            replacement = None
-            try:
-                assert forwarder.forward_request_admission_state_for_tests() == (2, 4)
-
-                with pytest.raises(forwarder.AuthBaseForwardingSaturatedError):
-                    await forwarder.forward_request(
-                        "https://example.com",
-                        "POST",
-                        [],
-                        b"12",
-                        admission=admission,
-                    )
-
-                assert forwarder.forward_request_admission_state_for_tests() == (1, 3)
-                replacement = forwarder.reserve_forward_request_admission(1)
-                assert forwarder.forward_request_admission_state_for_tests() == (2, 4)
-            finally:
-                forwarder.release_forward_request_admission(unrelated)
-                if replacement is not None:
-                    forwarder.release_forward_request_admission(replacement)
-
-        assert forwarder.forward_request_admission_state_for_tests() == (0, 0)
-        assert upstream.resolve_calls == []
-        assert upstream.sockets == []
-        assert upstream.connect_calls == []
-
-    async def test_closes_response_and_connection_on_success(self):
-        with fake_forwarder_upstream(headers=[("Content-Type", "application/json")]) as upstream:
-            status, body, _ = await forwarder.forward_request(
-                "https://example.com", "GET", [], None
-            )
-
-        assert status == 200
-        assert body == b"ok"
-        assert upstream.socket.response_file is not None
-        assert upstream.socket.response_file.closed
-        assert upstream.socket.closed
-
-    async def test_preserves_duplicate_headers_on_error_status(self):
-        with fake_forwarder_upstream(
-            status=429,
-            body=b"rate limited",
-            headers=[
-                ("WWW-Authenticate", "Bearer realm=one"),
-                ("WWW-Authenticate", "Bearer realm=two"),
-                ("Content-Type", "text/plain"),
-            ],
-        ) as upstream:
-            status, body, headers = await forwarder.forward_request(
-                "https://example.com", "GET", [], None
-            )
-
-        assert status == 429
-        assert body == b"rate limited"
-        assert headers.get_all("WWW-Authenticate") == ["Bearer realm=one", "Bearer realm=two"]
-        assert headers["Content-Type"] == "text/plain"
-        assert upstream.socket.response_file is not None
-        assert upstream.socket.response_file.closed
-        assert upstream.socket.closed
-
-    async def test_closes_response_when_read_raises(self):
-        with (
-            fake_forwarder_upstream(read_side_effect=OSError("socket closed")) as upstream,
-            pytest.raises(OSError, match="socket closed"),
-        ):
-            await forwarder.forward_request("https://example.com", "GET", [], None)
-
-        assert upstream.socket.response_file is not None
-        assert upstream.socket.response_file.closed
-        assert upstream.socket.closed
-
-    async def test_closes_connection_when_request_raises(self):
-        with (
-            fake_forwarder_upstream(send_side_effect=ConnectionError("connect failed")) as upstream,
-            pytest.raises(ConnectionError, match="connect failed"),
-        ):
-            await forwarder.forward_request("https://example.com", "GET", [], None)
-
-        assert upstream.socket.response_file is None
-        assert upstream.socket.closed
-
-    async def test_closes_connection_when_getresponse_raises(self):
-        with (
-            fake_forwarder_upstream(
-                socket_factory=lambda: FakeSocket(b"not an HTTP response\r\n")
-            ) as upstream,
-            pytest.raises(http_client.BadStatusLine),
-        ):
-            await forwarder.forward_request("https://example.com", "GET", [], None)
-
-        assert upstream.socket.response_file is not None
-        assert upstream.socket.response_file.closed
-        assert upstream.socket.closed
 
 
 class TestForwardRequestAbortHandle:
